requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
minio==7.2.0
clickhouse-driver==0.2.6
clickhouse-connect==0.6.23
streamlit==1.28.0
pandas==2.1.3
numpy==1.26.2
pyyaml==6.0.1
plotly==5.18.0
